serviceOrderMangerApp/requirements.txt
asgiref==3.7.2
bcrypt==3.2.0
blinker==1.4
boto3==1.33.8
botocore==1.33.8
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.12
django-cors-headers==4.3.1
django-filter==23.5
django-rest-authtoken==2.1.4
django-storages==1.14.2
djangorestframework==3.14.0
fasteners==0.14.1
future==0.18.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
python-dateutil==2.8.1
python-decouple==3.8
pytz==2022.1
pyxdg==0.27
reportlab==3.6.8
requests==2.25.1
s3transfer==0.8.2
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.8.0
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
wadllib==1.3.6
xdg==5
zipp==1.0.0
